Tweaks...

+++ install/instcomposer.php
@@ -5,22 +5,32 @@
 $status = '';
 $failure = false;
 
-try {
-    if( ($f = popen("curl --remote-name https://getcomposer.org/composer.phar 2>&1", "r")) ) {
-        while( !feof($f) ){
+if(file_exists("../composer.lock")){
 
-            $status = fread($f, 1024);
-            $arr_content = array("percent"=> 95, "message" => "Downloading composer.phar... <br>". $status, "failure" => 0);
-            file_put_contents("tmp/" . session_id() . ".txt", json_encode($arr_content));
-            flush(); // you have to flush buffer
+    unlink("../composer.lock");
+
+}
+
+if(!file_exists("composer.phar")){
+
+    try {
+        if( ($f = popen("curl --remote-name https://getcomposer.org/composer.phar 2>&1", "r")) ) {
+            while( !feof($f) ){
+
+                $status = fread($f, 1024);
+                $arr_content = array("percent"=> 95, "message" => "Downloading composer.phar... <br>". $status, "failure" => 0);
+                file_put_contents("tmp/" . session_id() . ".txt", json_encode($arr_content));
+                flush(); // you have to flush buffer
+            }
+            fclose($f);
         }
-        fclose($f);
-    }
-} catch (Exception $e) {
+    } catch (Exception $e) {
 
-    $arr_content = array("percent"=> 95, "message" => $e->getMessage(), "failure" => 1);
-    file_put_contents("tmp/" . session_id() . ".txt", json_encode($arr_content));
-    flush();
+        $arr_content = array("percent"=> 95, "message" => $e->getMessage(), "failure" => 1);
+        file_put_contents("tmp/" . session_id() . ".txt", json_encode($arr_content));
+        flush();
+    }
+
 }
 
 try {
